Allow passing of y-data to Tabulated1D, read_function
Useful for this feature because we have the modified q_recoverable
data, without local neutrons, but we can't add it to a proper
dataset. Instead, allow the construction of a Tabulated1D
isntance with the desired y-data. Remaining interpolation data,
abscissa, number of points, mode, etc. will be read from the dataset.

read_function accepts a pointer to an xarray and will use this
information to create a Tabulated1D object either with the
explicit y-data, or that in the database accordingly.