Clean up Tabulated1D construction w, w/o y-data
Rely on a common get_params function to read the dataset and
get number of interpolation points, scheme, etc.
If the y-data is not passed, both the x and y data are
read from the dataset, similar to before, in the original
construction.
If the y-data is passed, then only the x-data is read from
the dataset in the newly added constructor.
Both constructors pass the dataset, x-data, and y-data to
get_params for the remainder of the processing