Add classes for depleting w/o multiprocessing
Two classes in openmc/deplete/pool.py are introduced to
handle depleting with and without multiprocessing. Through a
discussion on the user's group
https://groups.google.com/forum/#!topic/openmc-users/xHKYV-EBgrY
it was determined that some computing environments don't support
calls to fork() that multiprocessing.Pool requires. This can
lead to the depletion hanging indefinitely.

The PoolShim class introduced here is a drop-in replacement
for multiprocessing.Pool, at least as it is necessary for depletion.
It provides __enter__, __exit__, and starmap methods. The latter
passes off to itertools.starmap.

The DepletionDispatcher is responsible for solving the Bateman
equations, with or without multiprocessing. It provides a deplete
method (to which pool.deplete falls back on), using
multiprocessing.Pool or PoolShim depending on a construction argument